refactor(web/applications): move case folder to root (boas-740)
