+++ apps/web/src/server/applications/applications.router.js
@@ -4,7 +4,7 @@ import * as controller from './applications.controller.js';
 import * as filters from './applications.filters.js';
 import * as guards from './applications.guards.js';
 import * as locals from './applications.locals.js';
-import applicationsCaseRouter from './pages/case/applications-case.router.js';
+import applicationsCaseRouter from './case/applications-case.router.js';
 import applicationsCreateRouter from './pages/create-new-case/applications-create.router.js';
 import applicationsSearchRouter from './search-results/applications-search.router.js';
