+++ apps/web/src/server/applications/case/documentation-metadata/documentation-metadata.router.js
@@ -1,5 +1,5 @@
 import { Router as createRouter } from 'express';
-import asyncRoute from '../../../../../lib/async-route.js';
+import asyncRoute from '../../../lib/async-route.js';
 import * as controller from './documentation-metadata.controller.js';
 import * as locals from './documentation-metadata.locals.js';
 import * as validators from './documentation-metadata.validators.js';
