+++ apps/web/src/server/applications/case/documentation/applications-documentation.controller.js
@@ -2,7 +2,7 @@ import { sortBy } from 'lodash-es';
 import {
 	getSessionFilesNumberOnList,
 	setSessionFilesNumberOnList
-} from '../../../common/services/session.service.js';
+} from '../../common/services/session.service.js';
 import {
 	deleteCaseDocumentationFile,
 	getCaseDocumentationFileInfo,
@@ -13,9 +13,9 @@ import {
 
 /** @typedef {import('@pins/express').ValidationErrors} ValidationErrors */
 /** @typedef {import('../applications-case.locals.js').ApplicationCaseLocals} ApplicationCaseLocals */
-/** @typedef {import('../../../applications.types').DocumentationCategory} DocumentationCategory */
-/** @typedef {import('../../../applications.types').DocumentationFile} DocumentationFile */
-/** @typedef {import('../../../common/services/session.service.js').SessionWithFilesNumberOnList} SessionWithFilesNumberOnList */
+/** @typedef {import('../../applications.types').DocumentationCategory} DocumentationCategory */
+/** @typedef {import('../../applications.types').DocumentationFile} DocumentationFile */
+/** @typedef {import('../../common/services/session.service.js').SessionWithFilesNumberOnList} SessionWithFilesNumberOnList */
 /** @typedef {import('./applications-documentation.types').CaseDocumentationUploadProps} CaseDocumentationUploadProps */
 /** @typedef {import('./applications-documentation.types').CaseDocumentationBody} CaseDocumentationBody */
 /** @typedef {import('./applications-documentation.types').CaseDocumentationProps} CaseDocumentationProps */
