+++ apps/web/src/server/applications/case/edit/applicant/applications-edit-applicant.router.js
@@ -1,8 +1,8 @@
 import { Router as createRouter } from 'express';
-import asyncRoute from '../../../../../lib/async-route.js';
-import { registerCaseWithQuery } from '../../../../applications.locals.js';
-import { registerApplicantId } from '../../../create-new-case/applicant/applications-create-applicant.locals.js';
-import * as validators from '../../../create-new-case/applicant/applications-create-applicant.validators.js';
+import asyncRoute from '../../../../lib/async-route.js';
+import { registerCaseWithQuery } from '../../../applications.locals.js';
+import { registerApplicantId } from '../../../pages/create-new-case/applicant/applications-create-applicant.locals.js';
+import * as validators from '../../../pages/create-new-case/applicant/applications-create-applicant.validators.js';
 import * as controller from './applications-edit-applicant.controller.js';
 
 const applicationsEditApplicantRouter = createRouter();
