+++ apps/web/src/server/applications/case/edit/case/__tests__/applications-edit-case.test.js
@@ -1,14 +1,14 @@
 import { parseHtml } from '@pins/platform';
 import nock from 'nock';
 import supertest from 'supertest';
-import { fixtureCases } from '../../../../../../../../testing/applications/fixtures/cases.js';
+import { fixtureCases } from '../../../../../../../testing/applications/fixtures/cases.js';
 import {
 	fixtureRegions,
 	fixtureSectors,
 	fixtureSubSectors,
 	fixtureZoomLevels
-} from '../../../../../../../../testing/applications/fixtures/options-item.js';
-import { createTestEnvironment } from '../../../../../../../../testing/index.js';
+} from '../../../../../../../testing/applications/fixtures/options-item.js';
+import { createTestEnvironment } from '../../../../../../../testing/index.js';
 
 const { app, installMockApi, teardown } = createTestEnvironment();
 const request = supertest(app);
